requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Werkzeug==0.9.4
beautifulsoup4==4.3.2
gunicorn==18.0
itsdangerous==0.23
python-dateutil==2.2
requests==2.2.1
six==1.5.2
wsgiref==0.1.2
